+++ src/main/scala/scodec/Encoder.scala
@@ -26,6 +26,18 @@ trait Encoder[-A] { self =>
   def encode(value: A): String \/ BitVector
 
   /**
+   * Encodes the specified value in to a bit vector, throwing an
+   * `IllegalArgumentException` if encoding fails.
+   *
+   * @param value value to encode
+   * @param return error or binary encoding of the value
+   * @throws IllegalArgumentException upon encoding failure
+   * @group primary
+   */
+  final def encodeValid(value: A): BitVector =
+    encode(value) valueOr { err => throw new IllegalArgumentException(err) }
+
+  /**
    * Converts this encoder to an `Encoder[B]` using the supplied `B => A`.
    * @group combinators
    */
@@ -48,11 +60,11 @@ trait Encoder[-A] { self =>
 /** Provides functions for working with encoders. */
 trait EncoderFunctions {
 
-  /** Encodes the specified value to a bit vector. */
-  final def encode[A](enc: Encoder[A], a: A): String \/ BitVector = enc encode a
-
   /** Encodes the specified value to a bit vector using an implicitly available encoder. */
-  final def encode[A: Encoder](a: A): String \/ BitVector = encode(Encoder[A], a)
+  final def encode[A: Encoder](a: A): String \/ BitVector = Encoder[A].encode(a)
+
+  /** Encodes the specified value to a bit vector using an implicitly available encoder or throws an `IllegalArgumentException` if encoding fails. */
+  final def encodeValid[A: Encoder](a: A): BitVector = Encoder[A].encodeValid(a)
 
   /** Encodes the specified values, one after the other, to a bit vector using the specified encoders. */
   final def encodeBoth[A, B](encA: Encoder[A], encB: Encoder[B])(a: A, b: B): String \/ BitVector = for {